refactor(solobase): drop legacy_* modules; rename to canonical names
After Phase 5 wired all four flows on top of cli::helpers, the legacy_*
modules from the absorbed solobase-cli are no longer called. Delete
legacy_build and legacy_serve; rename legacy_config to config.
