+++ crates/solobase/src/cli/mod.rs
@@ -1,20 +1,18 @@
-//! Unified CLI scaffolding.
+//! Unified CLI surface.
 //!
 //! `cli_args` + `mode` define the parser and mode/target detection.
-//! `flows` contains the four (mode × target) handlers (stubbed in Task 7,
-//! implemented in Phase 5). `server` + `server_config` carry the legacy
-//! native server-boot body, invoked today by the sealed × native flow.
-//! `legacy_*` + `cmd` + `skills` are the absorbed solobase-cli scaffolding,
-//! kept for the in-process tests under `tests/cli/` until Task 13 deletes
-//! them.
+//! `flows` contains the four (mode × target) handlers. `helpers` holds
+//! the cross-flow utilities (block discovery, frontend copy, overlay
+//! application, wasm resolution, static-file HTTP server). `config` is
+//! the `solobase.toml` schema + walk-up loader. `server` + `server_config`
+//! carry the in-process native server-boot body, invoked today by the
+//! sealed × native flow. `cmd` is the child-process runner used by the
+//! flows that shell out (cargo, wasm-pack, wafer).
 pub mod cli_args;
 pub mod cmd;
+pub mod config;
 pub mod flows;
 pub mod helpers;
-pub mod legacy_build;
-pub mod legacy_config;
-pub mod legacy_serve;
 pub mod mode;
 pub mod server;
 pub mod server_config;
-pub mod skills;
