--- crates/solobase/tests/cli/fixtures/gizza-ai-style/blocks/example/Cargo.toml
@@ -1,3 +1,0 @@
-[package]
-name = "example"
-version = "0.1.0"
