--- crates/solobase/tests/cli/fixtures/gizza-ai-style/site/gizza-app.js
@@ -1,1 +1,0 @@
-// fixture
