--- crates/solobase/tests/cli/fixtures/gizza-ai-style/site/index.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html><html><body>gizza</body></html>
