--- crates/solobase/tests/cli/fixtures/solobase-web-style/solobase.toml
@@ -1,4 +1,0 @@
-[app]
-name = "solobase-web"
-title = "Solobase"
-boot_redirect = "/b/system/"
