--- crates/solobase/tests/cli/main.rs
@@ -1,6 +1,0 @@
-//! Integration test entry point for the legacy CLI modules.
-//!
-//! Cargo treats `tests/cli/main.rs` as a single test crate so the
-//! sibling `integration_smoke.rs` file can live alongside fixtures
-//! under `tests/cli/`.
-mod integration_smoke;
